Show menu for geneVariant legend, with shape selection
